requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
opencv-python==4.5.3.56
Pillow==8.3.1
pyserial==3.5
requests==2.26.0
urllib3==1.26.6
